refactor(overworld): レビュー指摘に対応する
- MacroWindow.Cols を int から consts.Chunk へ揃え、Rows との型非対称と
  int(col)<win.Cols 等の変換を解消する
- 本番コードを検証しない TestTileKeyFormat・TestExploredTilesKeyConsistency を
  撤去する。fmt.Sprintf を自作 struct へ当てるだけでキー形式変更を検知できない
- DiscoveredChunks・toCell の除算に、帯が有効なら ChunkW/ChunkH は正で
  ゼロ除算しない不変条件をコメントで明示する

+++ internal/overworld/macro_view_test.go
@@ -11,7 +11,7 @@ func TestFullBandWindow_帯全体に東西の余白を足す(t *testing.T) {
 	t.Parallel()
 	win := FullBandWindow(10, 4, 3)
 	assert.Equal(t, 10-MacroMargin, win.OriginX, "窓左端は東進位置から余白ぶん西")
-	assert.Equal(t, int(4+2*MacroMargin), win.Cols, "列数は帯幅に東西の余白を足す")
+	assert.Equal(t, 4+2*MacroMargin, win.Cols, "列数は帯幅に東西の余白を足す")
 	assert.Equal(t, consts.Chunk(3), win.Rows, "行数は帯の高さ")
 }
